unittests for user-annotation-state manager
