--- constants.py
@@ -1,6 +1,0 @@
-"""
-Constants used throughout POTATO.
-"""
-import os
-
-POTATO_HOME = os.environ.get("POTATO_HOME")
