+++ potato/app.py
@@ -11,7 +11,7 @@ from potato.db_utils.models.user_annotation_state import UserAnnotationState
 
 from potato.db_utils.user_manager import UserManager
 from potato.db_utils.user_annotation_state_manager import UserAnnotationStateManager
-from constants import POTATO_HOME
+from potato.constants import POTATO_HOME
 
 def create_app(config):
     """ App factory """
