+++ tests/test_config.yaml
@@ -1,0 +1,92 @@
+{
+    "port": 9001,
+    "server_name": "potato annotator",
+    "annotation_task_name": "Tests",
+
+    # Potato will write the annotation file for all annotations to this
+    # directory, as well as per-annotator output files and state information
+    # necessary to restart annotation.
+    "output_annotation_dir": "example-projects/dialogue_analysis/annotation_output/",
+
+    # The output format for the all-annotator data. Allowed formats are:
+    # * jsonl
+    # * json (same output as jsonl)
+    # * csv
+    # * tsv
+    #
+    "output_annotation_format": "tsv",
+
+    # If annotators are using a codebook, this will be linked at the top to the
+    # instance for easy access
+    "annotation_codebook_url": "",
+    "data_files": [
+       "example-projects/dialogue_analysis/data_files/dialogue-example.jsonl"
+    ],
+    "item_properties": {
+        "id_key": "id",
+        "text_key": "text",
+        "context_key": "context"
+    },
+    "user_config": {
+      "allow_all_users": True,
+      "users": [  ],
+    },
+
+    #list_as_text is used when the input text is actually a list of texts, usually used for best-worst-scaling or dialogue analysis
+    "list_as_text": {
+      "text_list_prefix_type": 'none', #whether automatically insert a prefix for each line, currently supporting 'number', 'alphabet', 'number'
+      "horizontal": True,
+    },
+
+    # How many seconds do you want the annotators spend on each instance, after
+    # that, an alert will be sent per alert_time_each_instance seconds.
+    "alert_time_each_instance": 10000000,
+    "annotation_schemes": [
+        {
+          "annotation_type": "highlight",
+          "name": "certainty",
+          "description": "Highlight which phrases make the sentence more or less certain",
+          "labels": [
+            "certain", "uncertain"
+          ],
+
+          # If true, numbers [1-len(labels)] will be bound to each
+          # label. Highlight selection annotations with more than 10 are not supported
+          # with this simple keybinding and will need to use the full item
+          # specification to bind all labels to keys.
+          "sequential_key_binding": True,
+        },
+        {
+            "annotation_type": "radio",
+            "name": "sentiment",
+            "description": "What kind of sentiment does the given text hold?",
+            "labels": [
+               "positive", "neutral", "negative",
+            ],
+
+            # If true, numbers [1-len(labels)] will be bound to each
+            # label. Aannotations with more than 10 are not supported with this
+            # simple keybinding and will need to use the full item specification
+            # to bind all labels to keys.
+            "sequential_key_binding": True,                        
+        },       
+    ],
+
+    # The html that changes the visualiztation for your task. Change this file
+    # to influence the layout and description of your task. This is not a full
+    # HTML page, just the piece that does lays out your task's pieces
+    "html_layout": "templates/examples/plain_layout.html",
+
+    # The core UI files for Potato. You should not need to change these normally.
+    #
+    # Exceptions to this might include:
+    # 1) You want to add custom CSS/fonts to style your task
+    # 2) Your layout requires additional JS/assets to render
+    # 3) You want to support additional keybinding magic
+    #
+    "base_html_template": "templates/base_template.html",
+    "header_file": "templates/header.html",
+
+    # This is where the actual HTML files will be generated
+    "site_dir": "potato/templates/",
+}
